# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.CategoryGradeResponse:
    properties:
      category:
        type: string
      grade:
        type: string
      score:
        type: integer
    type: object
  internal_handlers.CheckFixStatusResponse:
    properties:
      account_id:
        type: string
      domain:
        type: string
      is_linked:
        type: boolean
      latest_grade:
        type: string
      latest_score:
        type: integer
      latest_verified_at:
        type: string
      linked_at:
        type: string
      verification:
        $ref: '#/definitions/internal_handlers.CheckFixVerificationResponse'
    type: object
  internal_handlers.CheckFixSubmissionResponse:
    properties:
      grade:
        type: string
      message:
        type: string
      passed:
        type: boolean
      verification:
        $ref: '#/definitions/internal_handlers.CheckFixVerificationResponse'
    type: object
  internal_handlers.CheckFixVerificationResponse:
    properties:
      category_grades:
        items:
          $ref: '#/definitions/internal_handlers.CategoryGradeResponse'
        type: array
      critical_findings:
        type: integer
      days_until_expiry:
        type: integer
      domain:
        type: string
      domain_match:
        type: boolean
      expires_at:
        type: string
      high_findings:
        type: integer
      id:
        type: string
      is_valid:
        type: boolean
      low_findings:
        type: integer
      medium_findings:
        type: integer
      overall_grade:
        type: string
      overall_score:
        type: integer
      report_date:
        type: string
      report_hash:
        type: string
      verified_at:
        type: string
      verified_domain:
        type: string
    type: object
...
  internal_handlers.CreateRequirementAPIRequest:
    properties:
      description:
        type: string
      due_date:
        type: string
      max_report_age_days:
        type: integer
      minimum_grade:
        type: string
      passing_score:
        type: integer
      priority:
        type: string
      questionnaire_id:
        type: string
      relationship_id:
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - relationship_id
    - title
    - type
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_handlers.LinkAccountRequest:
    properties:
      account_id:
        type: string
    required:
    - account_id
    type: object
...
  internal_handlers.PaginatedRequirementsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handlers.RequirementResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handlers.PaginatedSupplierRequirementsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handlers.SupplierRequirementResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handlers.RequirementResponse:
    properties:
      assigned_at:
        type: string
      company_id:
        type: string
      created_at:
        type: string
      days_until_due:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      is_overdue:
        type: boolean
      max_report_age_days:
        type: integer
      minimum_grade:
        type: string
      passing_score:
        type: integer
      priority:
        type: string
      questionnaire_id:
        type: string
      relationship_id:
        type: string
      status:
        type: string
      status_history:
        items:
          $ref: '#/definitions/internal_handlers.RequirementStatusChangeResp'
        type: array
      supplier_id:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.RequirementStatusChangeResp:
    properties:
      changed_at:
        type: string
      from_status:
        type: string
      reason:
        type: string
      to_status:
        type: string
    type: object
...
  internal_handlers.SubmitCheckFixRequest:
    properties:
      report_hash:
        type: string
    required:
    - report_hash
    type: object
...
  internal_handlers.SupplierRequirementResponse:
    properties:
      assigned_at:
        type: string
      company_id:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      days_until_due:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      is_overdue:
        type: boolean
      minimum_grade:
        type: string
      passing_score:
        type: integer
      priority:
        type: string
      questionnaire_id:
        type: string
      response_id:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  internal_handlers.VerifyReportRequest:
    properties:
      report_hash:
        type: string
    required:
    - report_hash
    type: object
host: localhost:8080
info:
  contact:
    email: support@nisfix.io
    name: NisFix Support
  description: B2B Supplier Security Portal API - Companies manage supplier security
    requirements through questionnaires and CheckFix reports
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NisFix Backend API
  version: "1.0"
paths:
...
  /requirements:
    get:
      consumes:
      - application/json
      description: Lists all requirements for the company
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by relationship
        in: query
        name: relationship_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PaginatedRequirementsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List requirements
      tags:
      - Requirements
    post:
      consumes:
      - application/json
      description: Creates a new requirement for a supplier
      parameters:
      - description: Create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateRequirementAPIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.RequirementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a requirement
      tags:
      - Requirements
...
  /requirements/{id}/checkfix:
    get:
      consumes:
      - application/json
      description: Gets the CheckFix verification details for a requirement (company
        view)
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.CheckFixVerificationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get CheckFix verification for a requirement
      tags:
      - CheckFix
...
  /supplier/checkfix/link:
    delete:
      consumes:
      - application/json
      description: Removes the CheckFix account link
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlink CheckFix account
      tags:
      - CheckFix
    post:
      consumes:
      - application/json
      description: Links the supplier's CheckFix account
      parameters:
      - description: Account to link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LinkAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.CheckFixStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link CheckFix account
      tags:
      - CheckFix
  /supplier/checkfix/status:
    get:
      consumes:
      - application/json
      description: Gets the current CheckFix integration status for the supplier
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.CheckFixStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get CheckFix status
      tags:
      - CheckFix
  /supplier/checkfix/verify:
    post:
      consumes:
      - application/json
      description: Verifies a CheckFix report and stores the verification
      parameters:
      - description: Report to verify
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.VerifyReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.CheckFixVerificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify a CheckFix report
      tags:
      - CheckFix
...
  /supplier/requirements:
    get:
      consumes:
      - application/json
      description: Lists all requirements assigned to this supplier
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PaginatedSupplierRequirementsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List requirements
      tags:
      - Supplier Portal
...
  /supplier/requirements/{id}/checkfix:
    post:
      consumes:
      - application/json
      description: Submits a CheckFix report verification as a response to a requirement
      parameters:
      - description: Requirement ID
        in: path
        name: id
        required: true
        type: string
      - description: CheckFix report to submit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SubmitCheckFixRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.CheckFixSubmissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit CheckFix verification for a requirement
      tags:
      - CheckFix
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"